requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
contextlib2==21.6.0
Django==5.2.5
django-ninja==1.4.3
django-ninja-extra==0.30.1
iniconfig==2.1.0
injector==0.22.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
